requirements.txt
python-telegram-bot==21.6
python-dotenv==1.0.1
playwright==1.46.0